Fix ember-cli deprecations
Ember-cli is throwing deprecations due to the package no longer being
configured properly
