+++ package.json
@@ -2,10 +2,6 @@
   "name": "broccoli-asset-rev",
   "version": "0.0.17",
   "description": "broccoli asset revisions (fingerprint)",
-  "main": "lib/asset-rev.js",
-  "ember-addon": {
-    "main": "lib/ember-cli-main.js"
-  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
